+++ src/components/nvi/table/utils/constants.ts
