+++ src/pages/messages/components/NviPublicationPointsTableRow.tsx
@@ -3,8 +3,8 @@ import { useState } from 'react';
 import { Link as RouterLink } from 'react-router';
 import { NviCandidateGlobalStatusEnum, NviCandidateStatusEnum } from '../../../api/searchApi';
 import { PercentageWithIcon } from '../../../components/_molecules/PercentageWithIcon';
-import { ALTERNATIVE_TEXT_INSTEAD_OF_ZERO } from '../../../components/nvi/table/constants';
-import { selfOrDescendantHasPointValues } from '../../../components/nvi/table/nvi-aggregations-helpers';
+import { ALTERNATIVE_TEXT_INSTEAD_OF_ZERO } from '../../../components/nvi/table/utils/constants';
+import { selfOrDescendantHasPointValues } from '../../../components/nvi/table/utils/nvi-aggregations-helpers';
 import { HorizontalBox } from '../../../components/styled/Wrappers';
 import { CenteredTableCell, TableNumberSkeleton } from '../../../styles/table-styles';
 import { NviInstitutionStatusResponse } from '../../../types/nvi.types';
